Fix UI Inconsistencies in Evaluation Results (#118)
