+++ app/components/evaluations/EvalRunCard.tsx
@@ -2,18 +2,14 @@
 
 import { useState } from "react";
 import { useRouter } from "next/navigation";
-import { colors } from "@/app/lib/colors";
-import {
-  EvalJob,
-  AssistantConfig,
-  getScoreObject,
-} from "@/app/components/types";
-import { getStatusColor, formatCostUSD } from "@/app/components/utils";
-import { timeAgo } from "@/app/lib/utils";
-import ConfigModal from "@/app/components/ConfigModal";
-import ScoreDisplay from "@/app/components/ScoreDisplay";
+import type { EvalJob, AssistantConfig } from "@/app/lib/types/evaluation";
+import { getScoreObject } from "@/app/lib/utils/evaluation";
+import { getStatusColor } from "@/app/components/utils";
+import { timeAgo, formatCostUSD } from "@/app/lib/utils";
+import { ConfigModal, InfoTooltip } from "@/app/components";
+import ScoreDisplay from "@/app/components/evaluations/ScoreDisplay";
 import CostIcon from "@/app/components/icons/evaluations/CostIcon";
-import InfoTooltip from "@/app/components/InfoTooltip";
+import DatabaseIcon from "@/app/components/icons/evaluations/DatabaseIcon";
 
 export interface EvalRunCardProps {
   job: EvalJob;
@@ -33,91 +29,55 @@ export default function EvalRunCard({
 
   return (
     <div
-      className="rounded-lg overflow-hidden"
-      style={{
-        backgroundColor: colors.bg.primary,
-        boxShadow: "0 1px 3px rgba(0, 0, 0, 0.04)",
-        borderLeft: `3px solid ${statusColor.border}`,
-      }}
+      className={`rounded-lg overflow-hidden bg-bg-primary shadow-sm border-l-3 ${statusColor.border}`}
     >
       <div className="px-5 py-4">
-        {/* Row 1: Run Name (left) | Status (right) */}
         <div className="flex items-start justify-between gap-4">
           <div className="min-w-0 flex-1">
-            <div
-              className="text-sm font-semibold truncate"
-              style={{ color: colors.text.primary }}
-            >
+            <div className="text-sm font-semibold truncate text-text-primary">
               {job.run_name}
             </div>
             {job.inserted_at && (
-              <div
-                className="text-xs mt-0.5"
-                style={{ color: colors.text.secondary }}
-              >
+              <div className="text-xs mt-0.5 text-text-secondary">
                 {timeAgo(job.inserted_at)}
               </div>
             )}
             {/* Error message (if failed) */}
             {job.error_message && (
-              <div
-                className="mt-2 text-xs break-words overflow-hidden"
-                style={{ color: "hsl(8, 86%, 40%)" }}
-              >
+              <div className="mt-2 text-xs wrap-break-word overflow-hidden text-status-error-text">
                 {job.error_message}
               </div>
             )}
           </div>
           <span
-            className="px-2.5 py-1 rounded text-xs font-semibold uppercase tracking-wide flex-shrink-0"
-            style={{ backgroundColor: statusColor.bg, color: statusColor.text }}
+            className={`px-2.5 py-1 rounded text-xs font-semibold uppercase tracking-wide shrink-0 ${statusColor.bg} ${statusColor.text}`}
           >
             {job.status}
           </span>
         </div>
 
-        {/* Row 2: Scores */}
         {scoreObj && (
           <div className="mt-3">
             <ScoreDisplay score={scoreObj} errorMessage={job.error_message} />
           </div>
         )}
 
-        {/* Row 3: Dataset + Config + Cost (left) | Actions (right) */}
         <div className="flex items-center justify-between gap-4 mt-3">
-          <div
-            className="flex items-center gap-3 text-xs"
-            style={{ color: colors.text.secondary }}
-          >
+          <div className="flex items-center gap-3 text-xs text-text-secondary">
             {job.dataset_name && (
               <span className="flex items-center gap-1.5">
-                <svg
-                  className="w-3.5 h-3.5 flex-shrink-0"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                  strokeWidth={2}
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M4 7v10c0 2 3.6 3 8 3s8-1 8-3V7M4 7c0 2 3.6 3 8 3s8-1 8-3M4 7c0-2 3.6-3 8-3s8 1 8 3M4 12c0 2 3.6 3 8 3s8-1 8-3"
-                  />
-                </svg>
+                <DatabaseIcon className="shrink-0" />
                 {job.dataset_name}
               </span>
             )}
             {job.assistant_id && assistantConfig?.name && (
-              <span
-                className="px-1.5 py-0.5 rounded"
-                style={{ backgroundColor: colors.bg.secondary }}
-              >
+              <span className="px-1.5 py-0.5 rounded bg-bg-secondary">
                 {assistantConfig.name}
               </span>
             )}
             {job.cost?.total_cost_usd != null && (
               <span className="flex items-center gap-1.5">
-                <CostIcon className="flex-shrink-0" />
+                <CostIcon className="shrink-0" />
                 {formatCostUSD(job.cost.total_cost_usd)}
                 <InfoTooltip
                   text={
@@ -144,30 +104,21 @@ export default function EvalRunCard({
               </span>
             )}
           </div>
-          <div className="flex items-center gap-3 flex-shrink-0">
+          <div className="flex items-center gap-3 shrink-0">
             <button
               onClick={() => setIsConfigModalOpen(true)}
-              className="px-3 py-1.5 rounded-lg text-xs font-medium border"
-              style={{
-                backgroundColor: "transparent",
-                borderColor: colors.border,
-                color: colors.text.primary,
-              }}
+              className="px-3 py-1.5 rounded-lg text-xs font-medium border border-border bg-transparent text-text-primary"
             >
               View Config
             </button>
             <button
               onClick={() => router.push(`/evaluations/${job.id}`)}
               disabled={!isCompleted}
-              className="px-3 py-1.5 rounded-lg text-xs font-medium border cursor-pointer disabled:cursor-not-allowed"
-              style={{
-                backgroundColor: "transparent",
-                borderColor: colors.border,
-                color: isCompleted
-                  ? colors.text.primary
-                  : colors.text.secondary,
-                opacity: isCompleted ? 1 : 0.5,
-              }}
+              className={`px-3 py-1.5 rounded-lg text-xs font-medium border border-border bg-transparent cursor-pointer disabled:cursor-not-allowed ${
+                isCompleted
+                  ? "text-text-primary opacity-100"
+                  : "text-text-secondary opacity-50"
+              }`}
             >
               View Results
             </button>
