+++ app/components/evaluations/EvaluationsTab.tsx
@@ -4,12 +4,13 @@ import { useState, useEffect, useCallback } from "react";
 import { apiFetch } from "@/app/lib/apiClient";
 import { colors } from "@/app/lib/colors";
 import { Dataset } from "@/app/lib/types/dataset";
-import { EvalJob, AssistantConfig } from "@/app/components/types";
+import { EvalJob, AssistantConfig } from "@/app/lib/types/evaluation";
 import ConfigSelector from "@/app/components/ConfigSelector";
 import Loader from "@/app/components/Loader";
 import EvalRunCard from "./EvalRunCard";
 import EvalDatasetDescription from "./EvalDatasetDescription";
 import { useAuth } from "@/app/lib/context/AuthContext";
+import { RefreshIcon } from "@/app/components/icons";
 
 type Tab = "datasets" | "evaluations";
 
@@ -390,23 +391,12 @@ export default function EvaluationsTab({
               <button
                 onClick={fetchEvaluations}
                 disabled={isLoading}
-                className="p-1.5 rounded"
-                style={{ color: colors.text.secondary }}
+                className="p-1.5 rounded text-text-secondary cursor-pointer"
                 aria-label="Refresh evaluations"
               >
-                <svg
-                  className={`w-4 h-4 ${isLoading ? "animate-spin" : ""}`}
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
-                  />
-                </svg>
+                <RefreshIcon
+                  className={`w-4 h-4 -scale-x-100 ${isLoading ? "animate-spin" : ""}`}
+                />
               </button>
             </div>
           </div>
@@ -418,14 +408,12 @@ export default function EvaluationsTab({
               boxShadow: "0 1px 3px rgba(0, 0, 0, 0.04)",
             }}
           >
-            {/* Loading */}
             {isLoading && evalJobs.length === 0 && (
               <div className="p-16">
                 <Loader size="md" message="Loading evaluation runs..." />
               </div>
             )}
 
-            {/* Error */}
             {error && (
               <div className="p-4">
                 <div
@@ -439,7 +427,6 @@ export default function EvaluationsTab({
               </div>
             )}
 
-            {/* Empty State */}
             {!isLoading && evalJobs.length === 0 && !error && (
               <div className="p-16 text-center">
                 <svg
@@ -469,7 +456,6 @@ export default function EvaluationsTab({
               </div>
             )}
 
-            {/* Runs List */}
             {evalJobs.length > 0 &&
               (() => {
                 const filteredJobs =
