+++ app/components/icons/common/RefreshIcon.tsx
@@ -13,11 +13,13 @@ export default function RefreshIcon({ className, style }: IconProps) {
       strokeWidth={2}
       style={style}
     >
-      <path
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
-      />
+      <g transform="scale(-1,1) translate(-24,0)">
+        <path
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
+        />
+      </g>
     </svg>
   );
 }
