+++ app/components/text-to-speech/EvaluationsTab.tsx
@@ -15,6 +15,7 @@ import { useAuth } from "@/app/lib/context/AuthContext";
 import { apiFetch } from "@/app/lib/apiClient";
 import Loader, { LoaderBox } from "@/app/components/Loader";
 import { getStatusColor } from "@/app/components/utils";
+import { RefreshIcon } from "@/app/components/icons";
 import AudioPlayerFromUrl from "./AudioPlayerFromUrl";
 import { useToast } from "@/app/components/Toast";
 
@@ -442,22 +443,11 @@ export default function EvaluationsTab({
                 <button
                   onClick={loadRuns}
                   disabled={isLoadingRuns}
-                  className="p-1.5 rounded"
-                  style={{ color: colors.text.secondary }}
+                  className="p-1.5 rounded text-text-secondary cursor-pointer"
                 >
-                  <svg
+                  <RefreshIcon
                     className={`w-4 h-4 ${isLoadingRuns ? "animate-spin" : ""}`}
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
-                    />
-                  </svg>
+                  />
                 </button>
               </div>
             )}
@@ -1134,39 +1124,24 @@ export default function EvaluationsTab({
                       return (
                         <div
                           key={run.id}
-                          className="rounded-lg overflow-hidden"
-                          style={{
-                            backgroundColor: colors.bg.primary,
-                            boxShadow: "0 1px 3px rgba(0, 0, 0, 0.04)",
-                            borderLeft: `3px solid ${statusColor.border}`,
-                          }}
+                          className={`rounded-lg overflow-hidden bg-bg-primary shadow-sm border-l-3 ${statusColor.border}`}
                         >
                           <div className="px-5 py-4">
                             {/* Row 1: Run Name + Status */}
                             <div className="flex items-start justify-between gap-4">
                               <div className="min-w-0 flex-1">
-                                <div
-                                  className="text-sm font-semibold truncate"
-                                  style={{ color: colors.text.primary }}
-                                >
+                                <div className="text-sm font-semibold truncate text-text-primary">
                                   {run.run_name}
                                 </div>
                                 {/* Error message */}
                                 {run.error_message && (
-                                  <div
-                                    className="mt-2 text-xs break-words overflow-hidden"
-                                    style={{ color: "hsl(8, 86%, 40%)" }}
-                                  >
+                                  <div className="mt-2 text-xs wrap-break-word overflow-hidden text-status-error-text">
                                     {run.error_message}
                                   </div>
                                 )}
                               </div>
                               <span
-                                className="px-2.5 py-1 rounded text-xs font-semibold uppercase tracking-wide flex-shrink-0"
-                                style={{
-                                  backgroundColor: statusColor.bg,
-                                  color: statusColor.text,
-                                }}
+                                className={`px-2.5 py-1 rounded text-xs font-semibold uppercase tracking-wide shrink-0 ${statusColor.bg} ${statusColor.text}`}
                               >
                                 {run.status}
                               </span>
